feat: better class async builders and reorg inside webhooks
