+++ src/utils/transfers.ts
@@ -1,28 +1,10 @@
-import { Database } from '../db/conn';
 import {
   REWARDS_COLLECTION,
-  TRANSACTION_STATUS,
   TRANSFERS_COLLECTION,
   USERS_COLLECTION,
 } from './constants';
-import {
-  getPatchWalletAccessToken,
-  getPatchWalletAddressFromTgId,
-  sendTokens,
-} from './patchwallet';
-import { addTrackSegment } from './segment';
-import axios, { AxiosError } from 'axios';
-import {
-  FLOWXO_NEW_TRANSACTION_WEBHOOK,
-  FLOWXO_WEBHOOK_API_KEY,
-} from '../../secrets';
-import { Db, Document, ObjectId, WithId } from 'mongodb';
+import { Db, ObjectId } from 'mongodb';
 import { formatDate } from './time';
-import {
-  PatchRawResult,
-  TransactionParams,
-  TransactionStatus,
-} from '../types/webhook.types';
 import {
   IncomingTransaction,
   OutgoingTransaction,
@@ -266,196 +248,4 @@ export async function getRewardLinkTxsUser(
         )?.userHandle || null,
     })),
   );
-}
-
-/**
- * Creates a transfer specific to Telegram based on the specified parameters.
- * @param params - The parameters required for the transfer.
- * @returns A promise resolving to a TransferTelegram instance or a boolean value.
- *          - If the TransferTelegram instance is successfully created and initialized, it's returned.
- *          - If initialization of the transfer's database fails, returns `false`.
- */
-export async function createTransferTelegram(
-  params: TransactionParams,
-): Promise<TransferTelegram | boolean> {
-  const transfer = new TransferTelegram(params);
-  return (await transfer.initializeTransferDatabase()) && transfer;
-}
-
-/**
- * Represents a Telegram transfer.
- */
-export class TransferTelegram {
-  /** The parameters required for the transaction. */
-  params: TransactionParams;
-
-  /** Indicates if the transfer is present in the database. */
-  isInDatabase: boolean = false;
-
-  /** Transaction details of the transfer. */
-  tx: WithId<Document> | null;
-
-  /** Current status of the transfer. */
-  status: TransactionStatus;
-
-  /** Wallet address of the recipient. */
-  recipientWallet?: string;
-
-  /** Transaction hash associated with the transfer. */
-  txHash?: string;
-
-  /** User operation hash. */
-  userOpHash?: string;
-
-  /** Database reference. */
-  db: Db | null;
-
-  /**
-   * Constructor for TransferTelegram class.
-   * @param params - The parameters required for the transfer.
-   */
-  constructor(params: TransactionParams) {
-    // Properties related to user and transaction details
-    this.params = params;
-
-    // Default values if not provided
-    this.isInDatabase = false;
-    this.tx = null;
-    this.status = TRANSACTION_STATUS.UNDEFINED;
-    this.recipientWallet = undefined;
-    this.txHash = undefined;
-    this.userOpHash = undefined;
-  }
-
-  /**
-   * Initializes the transfer object by connecting to the database and retrieving relevant information.
-   * @returns {Promise<boolean>} - True if initialization is successful, false otherwise.
-   */
-  async initializeTransferDatabase(): Promise<boolean> {
-    this.db = await Database.getInstance();
-    this.tx = await this.getTransferFromDatabase();
-
-    try {
-      this.recipientWallet = await getPatchWalletAddressFromTgId(
-        this.params.recipientTgId,
-      );
-    } catch (error) {
-      return false;
-    }
-
-    if (this.tx) {
-      this.isInDatabase = true;
-      this.status = this.tx.status;
-      this.userOpHash = this.tx.userOpHash;
-    } else {
-      await this.updateInDatabase(TRANSACTION_STATUS.PENDING, new Date());
-    }
-
-    return true;
-  }
-
-  /**
-   * Retrieves the transfer information from the database.
-   * @returns {Promise<WithId<Document>>} - The transfer information or null if not found.
-   */
-  async getTransferFromDatabase(): Promise<WithId<Document> | null> {
-    if (this.db)
-      return await this.db
-        .collection(TRANSFERS_COLLECTION)
-        .findOne({ eventId: this.params.eventId });
-    return null;
-  }
-
-  /**
-   * Updates the transfer information in the database.
-   * @param {TransactionStatus} status - The transaction status.
-   * @param {Date|null} date - The date of the transaction.
-   */
-  async updateInDatabase(
-    status: TransactionStatus,
-    date: Date | null,
-  ): Promise<void> {
-    await this.db?.collection(TRANSFERS_COLLECTION).updateOne(
-      { eventId: this.params.eventId },
-      {
-        $set: {
-          eventId: this.params.eventId,
-          chainId: this.params.chainId,
-          tokenSymbol: this.params.tokenSymbol,
-          tokenAddress: this.params.tokenAddress,
-          senderTgId: this.params.senderInformation?.userTelegramID,
-          senderWallet: this.params.senderInformation?.patchwallet,
-          senderName: this.params.senderInformation?.userName,
-          senderHandle: this.params.senderInformation?.userHandle,
-          recipientTgId: this.params.recipientTgId,
-          recipientWallet: this.recipientWallet,
-          tokenAmount: this.params.amount,
-          status: status,
-          ...(date ? { dateAdded: date } : {}),
-          transactionHash: this.txHash,
-          userOpHash: this.userOpHash,
-        },
-      },
-      { upsert: true },
-    );
-    console.log(
-      `[${this.params.eventId}] transaction from ${this.params.senderInformation?.userTelegramID} to ${this.params.recipientTgId} for ${this.params.amount} in MongoDB as ${status} with transaction hash : ${this.txHash}.`,
-    );
-  }
-
-  /**
-   * Saves transaction information to the Segment.
-   * @returns {Promise<void>} - The result of adding the transaction to the Segment.
-   */
-  async saveToSegment(): Promise<void> {
-    // Add transaction information to the Segment
-    await addTrackSegment({
-      ...this.params,
-      dateAdded: new Date(),
-      transactionHash: this.txHash || '',
-      recipientWallet: this.recipientWallet || '',
-    });
-  }
-
-  /**
-   * Saves transaction information to FlowXO.
-   * @returns {Promise<void>} - The result of sending the transaction to FlowXO.
-   */
-  async saveToFlowXO(): Promise<void> {
-    // Send transaction information to FlowXO
-    await axios.post(FLOWXO_NEW_TRANSACTION_WEBHOOK, {
-      senderResponsePath: this.params.senderInformation?.responsePath,
-      chainId: this.params.chainId,
-      tokenSymbol: this.params.tokenSymbol,
-      tokenAddress: this.params.tokenAddress,
-      senderTgId: this.params.senderInformation?.userTelegramID,
-      senderWallet: this.params.senderInformation?.patchwallet,
-      senderName: this.params.senderInformation?.userName,
-      senderHandle: this.params.senderInformation?.userHandle,
-      recipientTgId: this.params.recipientTgId,
-      recipientWallet: this.recipientWallet,
-      tokenAmount: this.params.amount,
-      transactionHash: this.txHash,
-      dateAdded: new Date(),
-      apiKey: FLOWXO_WEBHOOK_API_KEY,
-    });
-  }
-
-  /**
-   * Sends a transaction action, triggering PatchWallet.
-   * @returns Promise<axios.AxiosResponse<PatchRawResult, AxiosError>> - Promise resolving to an AxiosResponse object with PatchRawResult data or AxiosError on failure.
-   */
-  async sendTransactionAction(): Promise<
-    axios.AxiosResponse<PatchRawResult, AxiosError>
-  > {
-    return await sendTokens(
-      this.params.senderInformation?.userTelegramID,
-      this.recipientWallet || '',
-      this.params.amount,
-      await getPatchWalletAccessToken(),
-      this.params.delegatecall || 0,
-      this.params.tokenAddress,
-      this.params.chainId,
-    );
-  }
 }
